requirements.txt
Scrapy==2.11.0
playwright==1.40.0
requests==2.31.0
scrapy-fake-useragent==1.4.4
psycopg2-binary==2.9.9
pgvector==0.2.4
sentence-transformers==2.2.2
streamlit==1.30.0
tiktoken==0.5.2